+++ CHANGES.md
@@ -1,12 +1,22 @@
 # CHANGELOG
 
+## 0.12.2 (????-??-??)
+
+### Bug fixes
+
+-   Fix writing non-string object columns with arrow (#630).
+-   Fix writing empty string category columns with arrow fails (#621).
+-   Fix Time type columns being skipped with `use_arrow=False` (#617).
+-   Fix overwriting a corrupt fileGDB directory (#600).
+-   Fix attribute data being incorrectly written with KML driver (#650).
+
 ## 0.12.1 (2025-11-28)
 
 ### Bug fixes
 
--   Fix regression in reading date columns (#616)
+-   Fix regression in reading date columns (#616).
 -   Fix regression in `read_dataframe` when `use_arrow=True` and `columns` is used to filter
-    out columns of some specific types (#611)
+    out columns of some specific types (#611).
 
 ## 0.12.0 (2025-11-26)